PS-7504: Add ability to build tarball with Debug binaries (#4183)
* Improve codestyle of linking-script logic
* Work as should with debug binaries
* Update tests

+++ build-ps/build-binary.sh
@@ -29,6 +29,8 @@ CMAKE_BUILD_TYPE=''
 COMMON_FLAGS=''
 #
 TOKUDB_BACKUP_VERSION=''
+# enable asan
+ENABLE_ASAN=0
 #
 # Some programs that may be overriden
 TAR=${TAR:-tar}
@@ -56,7 +58,12 @@ do
         shift
         CMAKE_BUILD_TYPE='Debug'
         BUILD_COMMENT="${BUILD_COMMENT:-}-debug"
-        DEBUG_EXTRA="-DDEBUG_EXTNAME=OFF"
+        TARBALL_SUFFIX="-debug"
+        DEBUG_EXTRA="-DDEBUG_EXTNAME=OFF -DWITH_DEBUG=ON"
+        ;;
+    -a | --asan )
+        shift
+        ENABLE_ASAN=1
         ;;
     -v | --valgrind )
         shift
@@ -163,7 +170,7 @@ else
     REVISION=""
 fi
 PRODUCT_FULL="Percona-Server-$MYSQL_VERSION-$PERCONA_SERVER_VERSION"
-PRODUCT_FULL="$PRODUCT_FULL${BUILD_COMMENT:-}-$TAG$(uname -s)${DIST_NAME:-}.$TARGET${GLIBC_VER:-}"
+PRODUCT_FULL="$PRODUCT_FULL-$TAG$(uname -s)${DIST_NAME:-}.$TARGET${GLIBC_VER:-}${TARBALL_SUFFIX:-}"
 COMMENT="Percona Server (GPL), Release ${MYSQL_VERSION_EXTRA#-}"
 COMMENT="$COMMENT, Revision $REVISION${BUILD_COMMENT:-}"
 
@@ -172,12 +179,14 @@ export CC=${CC:-gcc}
 export CXX=${CXX:-g++}
 
 # If gcc >= 4.8 we can use ASAN in debug build but not if valgrind build also
-if [[ "$CMAKE_BUILD_TYPE" == "Debug" ]] && [[ "${CMAKE_OPTS:-}" != *WITH_VALGRIND=ON* ]]; then
-  GCC_VERSION=$(${CC} -dumpversion)
-  GT_VERSION=$(echo -e "4.8.0\n${GCC_VERSION}" | sort -t. -k1,1nr -k2,2nr -k3,3nr | head -1)
-  if [ "${GT_VERSION}" = "${GCC_VERSION}" ]; then
-    DEBUG_EXTRA="${DEBUG_EXTRA} -DWITH_ASAN=ON"
-  fi
+if [[ $ENABLE_ASAN -eq 1 ]]; then
+    if [[ "$CMAKE_BUILD_TYPE" == "Debug" ]] && [[ "${CMAKE_OPTS:-}" != *WITH_VALGRIND=ON* ]]; then
+        GCC_VERSION=$(${CC} -dumpversion)
+        GT_VERSION=$(echo -e "4.8.0\n${GCC_VERSION}" | sort -t. -k1,1nr -k2,2nr -k3,3nr | head -1)
+        if [ "${GT_VERSION}" = "${GCC_VERSION}" ]; then
+            DEBUG_EXTRA="${DEBUG_EXTRA} -DWITH_ASAN=ON"
+        fi
+    fi
 fi
 
 # TokuDB cmake flags
@@ -197,7 +206,6 @@ then
         CMAKE_OPTS="${CMAKE_OPTS:-} -DUSE_VALGRIND=ON"
     fi
 fi
-
 #
 # Attempt to remove any optimisation flags from the debug build
 # BLD-238 - bug1408232
@@ -292,7 +300,6 @@ fi
     DIRLIST="bin lib lib/private lib/plugin lib/mysqlrouter/plugin lib/mysqlrouter/private"
 
     LIBPATH=""
-    OVERRIDE=false
 
     function gather_libs {
         local elf_path=$1
@@ -333,13 +340,13 @@ fi
         # Set proper runpath for bins but check before doing anything
         local elf_path=$1
         local r_path=$2
+        local mode=${3:-}
         for elf in $(find ${elf_path} -maxdepth 1 -exec file {} \; | grep 'ELF ' | cut -d':' -f1); do
             echo "Checking LD_RUNPATH for ${elf}"
             if [[ -z $(patchelf --print-rpath ${elf}) ]]; then
                 echo "Changing RUNPATH for ${elf}"
                 patchelf --set-rpath ${r_path} ${elf}
-            fi
-            if [[ ! -z "${override}" ]] && [[ "${override}" == "true" ]]; then
+            elif [[ ! -z ${mode} ]] && [[ ${mode} == "override" ]]; then
                 echo "Overriding RUNPATH for ${elf}"
                 patchelf --set-rpath ${r_path} ${elf}
             fi
@@ -379,19 +386,23 @@ fi
             gather_libs ${DIR}
         done
         # Set proper runpath
-        export override=false
         set_runpath bin '$ORIGIN/../lib/private/'
-        set_runpath lib '$ORIGIN/private/'
-        set_runpath lib/plugin '$ORIGIN/../private/'
+        if [[ ${CMAKE_BUILD_TYPE} == "Debug" ]]; then
+            set_runpath lib '$ORIGIN/private/' override
+            set_runpath lib/plugin '$ORIGIN/../private/' override
+        else
+            set_runpath lib '$ORIGIN/private/'
+            set_runpath lib/plugin '$ORIGIN/../private/'
+        fi
         set_runpath lib/private '$ORIGIN'
         # LIBS MYSQLROUTER
-        unset override && export override=true && set_runpath lib/mysqlrouter/plugin '$ORIGIN/:$ORIGIN/../private/:$ORIGIN/../../private/'
-        unset override && export override=true && set_runpath lib/mysqlrouter/private '$ORIGIN/:$ORIGIN/../plugin/:$ORIGIN/../../private/'
+        set_runpath lib/mysqlrouter/plugin '$ORIGIN/:$ORIGIN/../private/:$ORIGIN/../../private/' override
+        set_runpath lib/mysqlrouter/private '$ORIGIN/:$ORIGIN/../plugin/:$ORIGIN/../../private/' override
         #  BINS MYSQLROUTER
-        unset override && export override=true && set_runpath bin/mysqlrouter_passwd '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/'
-        unset override && export override=true && set_runpath bin/mysqlrouter_plugin_info '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/'
-        unset override && export override=true && set_runpath bin/mysqlrouter '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/'
-        unset override && export override=true && set_runpath bin/mysqlrouter_keyring '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/'
+        set_runpath bin/mysqlrouter_passwd '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/' override
+        set_runpath bin/mysqlrouter_plugin_info '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/' override
+        set_runpath bin/mysqlrouter '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/' override
+        set_runpath bin/mysqlrouter_keyring '$ORIGIN/../lib/mysqlrouter/private/:$ORIGIN/../lib/mysqlrouter/plugin/:$ORIGIN/../lib/private/' override
         # Replace libs
         for DIR in ${DIRLIST}; do
             replace_libs ${DIR}
@@ -402,18 +413,22 @@ fi
         done
     }
 
-    mkdir $INSTALLDIR/usr/local/minimal
-    cp -r "$INSTALLDIR/usr/local/$PRODUCT_FULL" "$INSTALLDIR/usr/local/minimal/$PRODUCT_FULL-minimal"
+    if [[ $CMAKE_BUILD_TYPE != "Debug" ]]; then
+        mkdir $INSTALLDIR/usr/local/minimal
+        cp -r "$INSTALLDIR/usr/local/$PRODUCT_FULL" "$INSTALLDIR/usr/local/minimal/$PRODUCT_FULL-minimal"
+    fi
 
     # NORMAL TARBALL
     cd "$INSTALLDIR/usr/local/$PRODUCT_FULL"
     link
 
     # MIN TARBALL
-    cd "$INSTALLDIR/usr/local/minimal/$PRODUCT_FULL-minimal"
-    rm -rf mysql-test 2> /dev/null
-    find . -type f -exec file '{}' \; | grep ': ELF ' | cut -d':' -f1 | xargs strip --strip-unneeded
-    link
+    if [[ $CMAKE_BUILD_TYPE != "Debug" ]]; then
+        cd "$INSTALLDIR/usr/local/minimal/$PRODUCT_FULL-minimal"
+        rm -rf mysql-test 2> /dev/null
+        find . -type f -exec file '{}' \; | grep ': ELF ' | cut -d':' -f1 | xargs strip --strip-unneeded
+        link
+    fi
 )
 
 # Package the archive
@@ -423,9 +438,11 @@ fi
     find $PRODUCT_FULL -type f -name 'COPYING.AGPLv3' -delete
     $TAR --owner=0 --group=0 -czf "$WORKDIR_ABS/$PRODUCT_FULL.tar.gz" $PRODUCT_FULL
 
-    cd "$INSTALLDIR/usr/local/minimal/"
-    find $PRODUCT_FULL-minimal -type f -name 'COPYING.AGPLv3' -delete
-    $TAR --owner=0 --group=0 -czf "$WORKDIR_ABS/$PRODUCT_FULL-minimal.tar.gz" $PRODUCT_FULL-minimal
+    if [[ $CMAKE_BUILD_TYPE != "Debug" ]]; then
+        cd "$INSTALLDIR/usr/local/minimal/"
+        find $PRODUCT_FULL-minimal -type f -name 'COPYING.AGPLv3' -delete
+        $TAR --owner=0 --group=0 -czf "$WORKDIR_ABS/$PRODUCT_FULL-minimal.tar.gz" $PRODUCT_FULL-minimal
+    fi
 )
 
 # Clean up
